Richa_Mishra/milestone_1/requirements.txt
transformers
sentence-transformers
torch
numpy
faiss-cpu
pdfplumber
nltk
tqdm